[changes] Queue uniformity, query statement uniformity
